+++ tests/testthat/test-plotting_funs.R
@@ -37,6 +37,7 @@ test_that("Plotting of butterfly plots from epochs", {
 })
 
 test_that("Topoplots", {
+  skip_on_appveyor()
   vdiffr::expect_doppelganger("topoplot of epochs",
                               topoplot(demo_epochs))
   vdiffr::expect_doppelganger("topoplot of epochs 150-200ms",